requirements.txt
cherrypy==18.6.1
paste==3.5.0
flask==2.0.1
pyspark==3.1.2
apache-sedona==1.0.1
kafka-python==1.3.5
geopandas==0.9.0
h3==0.10.9
googlemaps==4.4.7
pygenetic==1.0.2